Fixing flaky test
